requirements.txt
asgiref==3.4.1
boto3==1.18.31
botocore==1.21.31
dj-database-url==0.5.0
Django==3.2.6
django-crispy-forms==1.12.0
django-filter==2.4.0
django-storages==1.11.1
fontawesome-free==5.15.4
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
